Removed explicit initialization of latest Run Report.
Initially, the app was manually initialized with the latest Run Report. This worked, but was a duplication of code, as subsequently selected Run Reports were created from different code. Furthermore, the initializing code had no error checks, so the app would never start if the initial Run Report was faulty. Now the initial Run Report and subsequent selections are loaded with the same code.