Refactored fixSeqWareTSV() as Run Report format changed
The Run Report is being upgraded to make it more machine readable. The
new report breaks throws an error due to changes in column naming. To
allow the reading of differently formatted Run Reports, a column_naming
file is provided, specifying the Run Report column names that are in use
and the canonical names that will be used in the app.