Refactored when and where data.table is melted and split
Initially, both steps were done as soon as the data.table was loaded.
This was not ideal, as it assumed all downstream functionality can deal
with that format. By keeping the data.table intact, the split function
is no longer necessary and the melting happens just before plotting.
This setup also allows saner ordering of plots.