The Sidebar now allows for vertical scrolling.
As more elements are added to the Sidebar, page scrolling would have
become an issue. As the whole page would scroll, the plot would move out
of the field of view while the user was picking options. This CSS hack
allows only the Sidebar to scroll while keeping the plot on screen.