TSV files are loaded from individual folders.
Initially, all TSV files were in the root folder. As it has become
necessary to load JSON as well, it makes more sense to have the SeqWare
run folder that contains both TSV and JSON files. The code in this
commit is compatible with both ways of storing the TSV files.