User can select libraries and selection can be copied to clipboard
This uses the event_data functionality of plotly+shiny. The main problem
was that it does not allow the same plotted data point to be clicked on
twice. As a user should be able to select and unselect the same data
point easily, this needed to be addressed. The simplest solution found
was to use shinyjs to inject custom JavaScript which reset the click
event to NULL. This allowed the user to select the same plotted point
multiple times.